✒ Update 23andMe file name
Capitalize "Me" - this is more correct.

Also, remove username. Rationale: Someday we may want to enable studies
having "anonymous" access to private data (i.e. not know the username for a
connected participant) - this reduces security concerns for them. To that end,
we should avoid embedding the username in the file name.

In the future, we may want to override this original "filename" in the S3 key
to download a more convenient file name that includes the username, e.g. when
downloading the file from a public profile (where username is not secret).
This isn't a high priority feature, though.