🔧 Avoid reprocessing 23andMe data
We want to avoid reprocessing the same data if it's not needed.

To determine this, we store a hash of the pre-processed file's path as
metadata in processed files. Check this to decide when reprocessing is
unnecessary (because the source file is the same).

The source files (uploaded by users) include a UUID in their path,
so any given upload should produce a unique hash.